Fix threading issues and make buffers larger
To run a CFReadStream on a separate thread you must make sure the run
loop is run in default mode and is running until the audio has stopped
streaming.

Apple's recommended 3 buffers is too small for http streaming so buffers
were uped to 16 and 4096 bytes each.  These numbers could be adjusted to
have less delay in the begining.