Thread release and play end of buffer
Change the run loop termination condition so when there is nothing to do it will exit - otherwise the thread sits there after the stream plays doing nothing, using up resources.
Removing self.isPlaying = NO from the TDAudioStreamEventEnd handling as that is just the end of the stream - the queue still has buffered data to play, so the stream is still playing. [self.audioQueue finish] instructs it to flush out the rest of the queued data and then stop.
Adding self.isPlaying = NO; to the error handler and audioQueueDidFinishPlaying, the latter of which is called when the queue runs out of data to play and the stream is really finished.
Making the notifications come from the self object rather than nil - makes no difference if you were listening to nil (ie all) but enables differentiation of the stream you are listening too.